+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## [unpublished]
 ### Added
 - `email` and `additionalContacts` in API reference json schema (#380)
+- Allow assigning contact `roles` using the cli (#383)
 ### Changed
 - Updated versions of external actions and MongoDB in GitHub workflows (#374)
 - Allow updating contact's email and href separately using the command line (#381)
+- Conftest file to reflect a patient's contact with email, roles and additionalContacts (#383)
 ### Fixed
 - Fixed typo on conda installation docs (#373)